Introduce state for InfluxDB configuration
This introduces a new influxdb.config state that will install the toml
Python module via pip and manage the InfluxDB configuration and defaults
file.

It allows users to use this formula for installing InfluxDB only without
forcing them to manage the configuration file also. If they do, however,
decide to manage the configuration file the formula will behave as
expected.